refactor: create a folder for each page
