+++ src/pages/login/LoginPage.ts
@@ -1,9 +1,9 @@
-import AuthState from "../features/auth/AuthState";
-import Router from "../app/router/router";
+import AuthState from "../../features/auth/AuthState";
+import Router from "../../app/router/router";
 
-import Component from "../shared/Component";
-import LoginForm from "../features/auth/LoginForm";
-import { h2 } from "../ui/tags";
+import Component from "../../shared/Component";
+import LoginForm from "../../features/auth/LoginForm";
+import { h2 } from "../../ui/tags";
 
 import styles from "./LoginPage.module.css";
